refactor(focus-lock): separate logic of FocusLock component
