+++ .changeset/cuddly-boats-search.md
@@ -1,0 +1,5 @@
+---
+"@yamada-ui/focus-lock": minor
+---
+
+separate logic and UI in component FocusLock
